Refactor cache purging code for optimizations
