+++ acp/main_module.php
@@ -33,6 +33,9 @@ class main_module
 	/** @var \phpbb\log\log $log */
 	protected $log;
 
+	/** @var \phpbb\mediaembed\cache\purge $purge_cache */
+	protected $purge_cache;
+
 	/** @var \phpbb\request\request $request */
 	protected $request;
 
@@ -72,6 +75,7 @@ class main_module
 		$this->config_text = $this->container->get('config_text');
 		$this->language    = $this->container->get('language');
 		$this->log         = $this->container->get('log');
+		$this->purge_cache = $this->container->get('phpbb.mediaembed.purge_cache');
 		$this->request     = $this->container->get('request');
 		$this->template    = $this->container->get('template');
 		$this->user        = $this->container->get('user');
@@ -233,27 +237,7 @@ class main_module
 	 */
 	protected function purge_mediaembed_cache()
 	{
-		try
-		{
-			$iterator = new \DirectoryIterator($this->cache->cache_dir);
-		}
-		catch (\Exception $e)
-		{
-			return;
-		}
-		foreach ($iterator as $fileInfo)
-		{
-			if ($fileInfo->isDot() || $fileInfo->isDir())
-			{
-				continue;
-			}
-
-			$filename = $fileInfo->getFilename();
-			if (strpos($filename, 'http.') === 0)
-			{
-				$this->cache->remove_file($fileInfo->getPathname());
-			}
-		}
+		$this->purge_cache->mediaembed_purge_cache();
 
 		$this->log->add('admin', $this->user->data['user_id'], $this->user->ip, 'LOG_PHPBB_MEDIA_EMBED_CACHE_PURGED');
