+++ language/ar/acp.php
@@ -28,11 +28,11 @@ $lang = array_merge($lang,[
 	'ACP_MEDIA_ALLOW_SIG'				=> 'التوقيعات',
 	'ACP_MEDIA_ALLOW_SIG_EXPLAIN'		=> 'السماح للعضو بإستخدام وسم الوسائط في التوقيع.',
 	'ACP_MEDIA_ENABLE_CACHE'			=> 'Enable Media Embed cache',
-	'ACP_MEDIA_ENABLE_CACHE_EXPLAIN'	=> 'In some cases, a simple URL does not provide all the information needed to embed a resource and the external content has to be downloaded, inspected and the information extracted. This only happens once at parsing time, but if the same text is parsed multiple times (e.g. when editing a text) a local copy of the external content can be saved in the cache for performance.',
+	'ACP_MEDIA_ENABLE_CACHE_EXPLAIN'	=> 'In some case you may notice slower than normal performance when loading media from other sites, especially while loading the same content multiple times (e.g. when editing a post). Enabling this will cache the information Media Embed gathers from sites locally and should improve performance.',
 	'ACP_MEDIA_PARSE_URLS'				=> 'تحويل الروابط العادية',
 	'ACP_MEDIA_PARSE_URLS_EXPLAIN'		=> 'اختار “نعم” لتحويل الروابط العادية (التي لم تندرج بين الوسوم <samp>[media]</samp> أو <samp>[url]</samp>) إلى محتوى الوسائط. مع الملاحظة بأن تطبيق هذا الخيار سيكون على المشاركات الجديدة فقط, ولن تؤثر على المشاركات القديمة.',
 	'ACP_MEDIA_PURGE_CACHE'				=> 'Purge Media Embed cache',
-	'ACP_MEDIA_PURGE_CACHE_EXPLAIN'		=> 'Media Embed cache is being purged by cron once per day (if enabled above). The cache can be manially purged here.',
+	'ACP_MEDIA_PURGE_CACHE_EXPLAIN'		=> 'Media Embed cache is automatically purged once per day, however this button can be used to manually purge its cache now.',
 	'ACP_MEDIA_SITE_TITLE'				=> 'الموقع: %s',
 	'ACP_MEDIA_SITE_DISABLED'			=> 'هذا الموقع يتعارض مع BBCode آخر موجود: [%s]',
